Move hash_file() to utilities module
